refactor(audit): split enqueue helpers out of auditContextWrappers
Workers crashed at startup after c804cace because auditContextWrappers.ts
pulled in next/headers, which isn't present in the workers Docker image.
The hotfix used a dynamic import as a band-aid; this split addresses the
underlying coupling.

- New lib/auditContextEnqueue.ts: runtime-agnostic BullMQ helpers
  (enqueueWithAuditContext, ActorContextJobData, EnqueueSystemOptions),
  zero next/* imports, safe to load in any Node context.
- auditContextWrappers.ts: now Next.js-only (withAuditContext,
  withActionAuditContext, enrichFromApiAuth) with next/headers back at
  the top level — the dynamic-import workaround is no longer needed.
- Worker files, SyncService, scripts, and mixed-import app routes now
  import enqueue helpers from the new module.
- Enqueue tests moved to auditContextEnqueue.test.ts (no next/headers
  mock) so test suite proves the module loads cleanly in a Next-free
  environment.

Verified: grep -rn "next/" workers/*.ts returns zero hits for non-test
worker files. The workers image can now load auditContextEnqueue.ts
without any transitive Next.js requirement.
