+++ testplanit/scripts/trigger-forecast-recalc.ts
@@ -1,4 +1,4 @@
-import { enqueueWithAuditContext } from "../lib/auditContextWrappers";
+import { enqueueWithAuditContext } from "../lib/auditContextEnqueue";
 import { getAllTenantIds, isMultiTenantMode } from "../lib/multiTenantPrisma";
 import { getForecastQueue } from "../lib/queues";
 import { JOB_UPDATE_ALL_CASES } from "../workers/forecastWorker";
